ドライバーランキング、コンストラクターランキングの同期処理を実装

+++ core/models/constructor_standing.py
@@ -4,7 +4,7 @@ from django.db import models
 class ConstructorStandingModel(models.Model):
     constructor = models.CharField(max_length=50)
     position = models.IntegerField()
-    points = models.IntegerField()
+    points = models.FloatField()
     season = models.IntegerField()
 
     def __str__(self) -> str:
